+++ docs/sphinx/index.rst
@@ -1,4 +1,4 @@
-.. psychic-octo-robot documentation master file, created by
+.. gitlit documentation master file, created by
    sphinx-quickstart on Sun Sep 21 13:31:07 2014.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
